+++ .terserrc
@@ -1,0 +1,4 @@
+{
+    keep_classnames: true,
+    keep_fnames: true
+}
