+++ package.json
@@ -1,42 +1,72 @@
 {
-  "name": "cryptovoxels",
+  "name": "voxels-scripting-bundle",
   "version": "0.9.0",
   "description": "",
-  "main": "index.ts",
+  "engines": {
+    "node": ">= 14"
+  },
+  "iframe": "dist/iframe_scripting_host.js",
+  "hosted": "dist/hosted_scripting_host.js",
+  "bundle": "dist/voxels_scripting_bundle.js",
+  "targets": {
+    "iframe": {
+      "source": "src/index.ts",
+      "outputFormat": "global",
+      "engines": {
+        "browsers": "chrome 80"
+      }
+    },
+    "hosted": {
+      "source": "src/index.ts",
+      "outputFormat": "global",
+      "engines": {
+        "browsers": "> 0.5%, last 2 versions, not dead"
+      }
+    },
+    "bundle": {
+      "source": "src/index.ts",
+      "outputFormat": "esmodule",
+      "engines": {
+        "node": ">= 12"
+      }
+    }
+  },
   "scripts": {
-    "build": "webpack --config webpack.config.js",
-    "copy:cryptovoxels": "cp ./dist/scripting-host.js ../cryptovoxels/dist/scripting-host.js",
-    "copy:grid": "cp ./dist/scripting-host.js ../grid/scripting-host.js",
-    "build:min": "webpack --config webpack-min.config.js",
-    "test": "ts-mocha -r ts-node/register ./tests/**/*.test.ts",
+    "build": "npm run build:iframe && npm run build:hosted",
+    "build:iframe": "npx parcel build --target iframe --no-scope-hoist",
+    "build:hosted": "npx parcel build --target hosted --no-optimize --no-scope-hoist",
+    "build:module": "npx parcel build --target bundle --no-optimize --no-scope-hoist",
+    "build:module:types": "tsc src/index.ts --target es5 --declaration --skipLibCheck --downlevelIteration --excludeDirectories node_modules --outDir ./dist --types @babylonjs/core,node --lib es2019,DOM",
+    "copy:voxels": "cp ./dist/iframe_scripting_host.js ../cryptovoxels/dist/scripting-host.js",
+    "copy:hosted": "cp ./dist/hosted_scripting_host.js ../grid/scripting-host.js",
+    "test": "ts-mocha -r ts-node/register ./tests/**/*.test.ts --config tsconfig.json",
     "coverage": "nyc -r lcov -e .ts -x \"*.test.ts\" npm run test"
   },
   "author": "",
   "license": "ISC",
   "dependencies": {
-    "@babylonjs/core": "^4.0.3",
+    "@babylonjs/core": "^4.2.0",
     "lodash.throttle": "^4.1.1",
     "ndarray": "1.0.18",
     "node-fetch": "^2.6.1",
+    "parcel": "^2.5.0",
     "tinyify": "^2.5.1",
     "uuid": "3.3.2",
     "ws": "7.4.6"
   },
   "devDependencies": {
     "@types/chai": "^4.3.1",
-    "@types/mocha": "^9.1.0",
-    "@types/node": "^17.0.24",
+    "@types/jest": "^27.5.1",
+    "@types/mocha": "^9.1.1",
+    "@types/node": "^17.0.33",
     "chai": "^4.3.6",
+    "events": "^3.3.0",
     "mocha": "^9.2.2",
     "nyc": "^15.1.0",
     "tape": "4.9.1",
-    "terser-webpack-plugin": "^5.2.0",
-    "ts-loader": "^9.2.8",
     "ts-mocha": "^9.0.2",
     "ts-node": "^10.7.0",
     "typedoc": "^0.22.15",
-    "typescript": "^4.6.3",
-    "webpack": "^5.50.0",
-    "webpack-cli": "^4.8.0"
+    "typescript": "^4.6.3"
   }
 }
